+++ src/com/intellij/idea/plugin/hybris/impex/actions/ImpExValidateAction.kt
@@ -43,10 +43,8 @@ class ImpExValidateAction : ExecuteStatementAction<HybrisImpexConsole>(
             executionMode = ExecutionMode.VALIDATE
         )
 
-        console.isEditable = false
         project.service<ImpExExecutionClient>().execute(context) { coroutineScope, result ->
             console.print(result)
-            console.isEditable = true
         }
     }