+++ src/com/intellij/idea/plugin/hybris/tools/remote/console/actions/ConsoleImpExValidateAction.kt
@@ -30,7 +30,6 @@ import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.components.service
 import com.intellij.ui.AnimatedIcon
-import com.intellij.util.asSafely
 
 class ConsoleImpExValidateAction : AnAction(
     "Validate ImpEx",
@@ -42,21 +41,21 @@ class ConsoleImpExValidateAction : AnAction(
 
     override fun actionPerformed(e: AnActionEvent) {
         val project = e.project ?: return
-        val consoleService = project.service<HybrisConsoleService>()
-        val console = consoleService.getActiveConsole()
-            ?.asSafely<HybrisImpexConsole>() ?: return
+        val console = e.project
+            ?.service<HybrisConsoleService>()
+            ?.getActiveConsole()
+            ?: return
 
         val context = ImpExExecutionContext(
             content = console.content,
             executionMode = ExecutionMode.VALIDATE,
         )
 
-        console.isEditable = false
-
-        project.service<ImpExExecutionClient>().execute(context) { coroutineScope, result ->
-            console.print(result)
-            console.isEditable = true
-        }
+        project.service<ImpExExecutionClient>().execute(
+            context = context,
+            beforeCallback = { coroutineScope -> console.beforeExecution() },
+            resultCallback = { coroutineScope, result -> console.print(result) }
+        )
     }
 
     override fun update(e: AnActionEvent) {